+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - *fit.test_ks*
 
 ## Modified functions:
+13/06/2024
+Modified dtype of empty arrays in Spectra
 04/04/2024
 - *Spectrum_2D.__init__*: reshape of the fid is now made with last dimension implicit
 02/05/2024